# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Psych Manager Origin API",
        "contact": {
            "email": "info@ps.ru"
        },
        "version": "1.0 beta"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "PMO server"
        }
    ],
    "paths": {
...
        "/sessions": {
            "get": {
                "tags": [
                    "Сессии"
                ],
                "summary": "Index page",
                "operationId": "sessions-index",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionIndexResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Список пуст"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Сессии"
                ],
                "summary": "Store",
                "operationId": "sessions-store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Клиент добавлен"
                    },
                    "500": {
                        "description": "Ошибка создания"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            }
        },
        "/sessions/{id}": {
            "get": {
                "tags": [
                    "Сессии"
                ],
                "summary": "Show page",
                "operationId": "sessions-show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id сессии",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionShowResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Сессии"
                ],
                "summary": "Update",
                "operationId": "sessions-update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/SessionRequest"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Клиент обновлён"
                    },
                    "404": {
                        "description": "Клиент не найден"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Сессии"
                ],
                "summary": "Delete",
                "operationId": "sessions-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id сессии",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cancel_reason": {
                                        "description": "Причина отмены",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запись отменена"
                    },
                    "404": {
                        "description": "Запись не найдена"
                    },
                    "500": {
                        "description": "Ошибка отмены"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            }
        },
        "/static-data/get-connection-type": {
            "get": {
                "tags": [
                    "Статические данные"
                ],
                "summary": "Connection types",
                "operationId": "get-connection-type",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Список пуст"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            }
        },
        "/static-data/get-gender-list": {
            "get": {
                "tags": [
                    "Статические данные"
                ],
                "summary": "Gender list",
                "operationId": "get-gender-list",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Список пуст"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SessionRequest": {
                "properties": {
                    "client_id": {
                        "description": "Class SessionRequest",
                        "type": "integer"
                    },
                    "doctor_id": {
                        "description": "Class SessionRequest",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Class SessionRequest",
                        "type": "string"
                    },
                    "session_date": {
                        "description": "date_format Y-m-d H:i:s",
                        "type": "string"
                    },
                    "comment": {
                        "description": "План на сессию",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SessionIndexResource": {
                "properties": {
                    "id": {
                        "description": "Class IndexResource",
                        "type": "string"
                    },
                    "name": {
                        "description": "Class IndexResource",
                        "type": "string"
                    },
                    "email": {
                        "description": "Class IndexResource",
                        "type": "string"
                    },
                    "connection_type": {
                        "description": "id записи типа связи",
                        "type": "integer"
                    },
                    "connection_type_string": {
                        "description": "Имя записи типа связи",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SessionShowResource": {
                "properties": {
                    "id": {
                        "description": "Class ShowResource",
                        "type": "integer"
                    },
                    "client_id": {
                        "description": "Class ShowResource",
                        "type": "integer"
                    },
                    "client_name": {
                        "description": "Class ShowResource",
                        "type": "string"
                    },
                    "doctor_id": {
                        "description": "Class ShowResource",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Class ShowResource",
                        "type": "string"
                    },
                    "comment": {
                        "description": "Комментарий/план на сессию",
                        "type": "string"
                    },
                    "session_date": {
                        "description": "Дата встречи",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "X-API-KEY": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}